WorkingCamera/requirements.txt
# ESP32 Camera Controller Dependencies
requests>=2.28.0
opencv-python>=4.7.0
Pillow>=9.0.0
numpy>=1.21.0
